hisaboalabackend-alamin/requirements.txt
asgiref==3.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.4
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-rest-knox==5.0.2
djangorestframework==3.16.0
drf-yasg==1.21.10
dynaconf==3.2.11
hvac==2.3.0
idna==3.10
inflection==0.5.1
knox==0.1.14
loguru==0.7.3
packaging==25.0
pillow==11.3.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
shortuuid==1.0.13
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
win32_setctime==1.2.0
